# faster way to get cpu count which takes .003ms vs .014ms