--- a/AppController/lib/datastore_server.rb
+++ b/AppController/lib/datastore_server.rb
@@ -66,7 +66,7 @@ module DatastoreServer
   # Number of servers is based on the number of CPUs.
   def self.number_of_servers
     # If this is NaN then it returns 0
-    num_procs = `cat /proc/cpuinfo | grep processor | wc -l`.to_i
+    num_procs = `grep -c processor /proc/cpuinfo`.to_i
     return DEFAULT_NUM_SERVERS if num_procs.zero?
     servers = num_procs * MULTIPLIER
     return 1 if servers.zero?