Stumbled across this, not sure if we really need the change. But its faster to just use grep on the file and get the count directly via the grep command:

```
1514% time cat /proc/cpuinfo | grep processor | wc -l
4
cat /proc/cpuinfo  0.00s user 0.00s system 48% cpu 0.004 total
grep processor  0.00s user 0.00s system 26% cpu 0.006 total
wc -l  0.00s user 0.00s system 39% cpu 0.004 total


1515% time grep -c processor /proc/cpuinfo
4
grep -c processor /proc/cpuinfo  0.00s user 0.00s system 65% cpu 0.003 total

```