@scragraham 2018-05-18
Well it looks like we don't use this code anymore anyways as the CPU count is gathered via Hermes (which gets it via psutil). This PR is no longer valid as the code is now gone in master.